+++ src/types/events.ts
@@ -1,0 +1,14 @@
+import {BigNumber} from 'ethers';
+
+export type LogMint = {
+  minter: string;
+  blockId: BigNumber;
+  staked: BigNumber;
+  power: BigNumber;
+};
+
+export type LogVoter = {
+  blockId: BigNumber;
+  voter: string;
+  vote: BigNumber;
+};