+++ test/services/BlockMinter.test.ts
@@ -23,7 +23,7 @@ import Block from '../../src/models/Block';
 import {leafWithAffidavit} from '../fixtures/leafWithAffidavit';
 import {loadTestEnv} from '../helpers/loadTestEnv';
 import TimeService from '../../src/services/TimeService';
-import {generateAffidavit, recoverSigner, signAffidavitWithWallet, sortLeaves} from '../../src/utils/mining';
+import {generateAffidavit, recoverSigner, signAffidavitWithWallet, sortLeaves, timestamp} from '../../src/utils/mining';
 
 describe('BlockMinter', () => {
   let mockedBlockchain: sinon.SinonStubbedInstance<Blockchain>;
@@ -88,10 +88,10 @@ describe('BlockMinter', () => {
   describe('#sortLeaves', () => {
     it("sorts leaves based on leaf's label", async () => {
       const leaves: Leaf[] = [
-        {label: 'b', _id: '1', timestamp: new Date(), blockHeight: 1, valueBytes: '0x0'},
-        {label: 'a', _id: '1', timestamp: new Date(), blockHeight: 1, valueBytes: '0x0'},
-        {label: 'd', _id: '1', timestamp: new Date(), blockHeight: 1, valueBytes: '0x0'},
-        {label: 'c', _id: '1', timestamp: new Date(), blockHeight: 1, valueBytes: '0x0'},
+        {label: 'b', _id: '1', timestamp: new Date(), blockId: 1, valueBytes: '0x0'},
+        {label: 'a', _id: '1', timestamp: new Date(), blockId: 1, valueBytes: '0x0'},
+        {label: 'd', _id: '1', timestamp: new Date(), blockId: 1, valueBytes: '0x0'},
+        {label: 'c', _id: '1', timestamp: new Date(), blockId: 1, valueBytes: '0x0'},
       ];
       const resultingLeaves = sortLeaves(leaves);
 
@@ -107,7 +107,7 @@ describe('BlockMinter', () => {
 
   describe('#generateAffidavit', () => {
     it('generates affidavit successfully', () => {
-      const affidavit = generateAffidavit(1, ethers.utils.keccak256('0x1234'), 1, ['ETH-USD'], [100]);
+      const affidavit = generateAffidavit(1, ethers.utils.keccak256('0x1234'), ['ETH-USD'], [100]);
 
       expect(affidavit)
         .to.be.a('string')
@@ -148,12 +148,13 @@ describe('BlockMinter', () => {
         '0x123',
         {
           blockNumber: BigNumber.from(1),
-          lastBlockHeight: BigNumber.from(0),
-          nextBlockHeight: BigNumber.from(1),
+          timePadding: 10,
+          lastBlockId: 1,
+          nextBlockId: 2,
           nextLeader: Wallet.createRandom().address,
-          validators: [wallet.address],
+          validators: [wallet.address, 'leader'],
           locations: ['abc'],
-          lastDataTimestamp: BigNumber.from(1),
+          lastDataTimestamp: timestamp(),
           powers: [BigNumber.from(1)],
           staked: BigNumber.from(1),
         },
@@ -172,12 +173,13 @@ describe('BlockMinter', () => {
         '0x123',
         {
           blockNumber: BigNumber.from(1),
-          lastBlockHeight: BigNumber.from(1),
-          nextBlockHeight: BigNumber.from(1),
+          timePadding: 100,
+          lastBlockId: 1,
+          nextBlockId: 1,
           nextLeader: wallet.address,
           validators: [wallet.address],
           locations: ['abc'],
-          lastDataTimestamp: BigNumber.from(1),
+          lastDataTimestamp: timestamp(),
           powers: [BigNumber.from(1)],
           staked: BigNumber.from(1),
         },
@@ -201,12 +203,13 @@ describe('BlockMinter', () => {
         '0x123',
         {
           blockNumber: BigNumber.from(1),
-          lastBlockHeight: BigNumber.from(0),
-          nextBlockHeight: BigNumber.from(1),
+          timePadding: 1,
+          lastBlockId: 0,
+          nextBlockId: 1,
           nextLeader: wallet.address,
           validators: [wallet.address],
           locations: ['abc'],
-          lastDataTimestamp: BigNumber.from(1),
+          lastDataTimestamp: 1,
           powers: [BigNumber.from(1)],
           staked: BigNumber.from(1),
         },
@@ -223,7 +226,6 @@ describe('BlockMinter', () => {
       expect(mockedSignatureCollector.apply.args[0][0]).to.be.deep.eq(
         {
           dataTimestamp: 10,
-          blockHeight: 1,
           fcd: fcd,
           leaves: fcd,
           signature,
@@ -250,12 +252,13 @@ describe('BlockMinter', () => {
         '0x123',
         {
           blockNumber: BigNumber.from(1),
-          lastBlockHeight: BigNumber.from(0),
-          nextBlockHeight: BigNumber.from(1),
+          timePadding: 1,
+          lastBlockId: 0,
+          nextBlockId: 1,
           nextLeader: wallet.address,
           validators: [wallet.address],
           locations: ['abc'],
-          lastDataTimestamp: BigNumber.from(1),
+          lastDataTimestamp: 1,
           powers: [BigNumber.from(1)],
           staked: BigNumber.from(1),
         },
@@ -291,12 +294,13 @@ describe('BlockMinter', () => {
         '0x123',
         {
           blockNumber: BigNumber.from(1),
-          lastBlockHeight: BigNumber.from(0),
-          nextBlockHeight: BigNumber.from(1),
+          timePadding: 0,
+          lastBlockId: 1,
+          nextBlockId: 1,
           nextLeader: wallet.address,
           validators: [wallet.address],
           locations: ['abc'],
-          lastDataTimestamp: BigNumber.from(1),
+          lastDataTimestamp: 1,
           powers: [BigNumber.from(1)],
           staked: BigNumber.from(1),
         },
@@ -312,7 +316,27 @@ describe('BlockMinter', () => {
         {signature, power: BigNumber.from(1), discrepancies: [], version: '1.0.0'},
       ]);
       mockedChainContract.submit.resolves({
-        wait: () => Promise.resolve({status: 1, transactionHash: '123'}),
+        wait: () =>
+          Promise.resolve({
+            status: 1,
+            transactionHash: '123',
+            logs: [
+              {
+                transactionIndex: 0,
+                blockNumber: 6618,
+                transactionHash: '0x17063b26e48f5d9862688aac0ce693e2dfc4d8d9f230573c331e6616d7a85b55',
+                address: '0xc4905364b78a742ccce7B890A89514061E47068D',
+                topics: [
+                  '0x5f11830295067c4bcc7d02d4e3b048cd7427be50a3aeb6afc9d3d559ee64bcfa',
+                  '0x000000000000000000000000998cb7821e605cc16b6174e7c50e19adb2dd2fb0',
+                ],
+                data:
+                  '0x000000000000000000000000000000000000000000000000000000000000033f00000000000000000000000000000000000000000000000029a2241af62c00000000000000000000000000000000000000000000000000001bc16d674ec80000',
+                logIndex: 1,
+                blockHash: '0x7422c3bf9cda4cd91e282a495945d4b4ff310a06a67614e806bf6bb244527225',
+              },
+            ],
+          }),
       } as any); // throw error when trying to submit minted block
 
       await blockMinter.apply();